backend/requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
pytz==2019.2
selenium==3.141.0
sqlparse==0.3.0
urllib3==1.25.3
